Made wiki as a separate module
